UIDATIMP-1151 Fix unit tests for the ViewAllLogs component and increase test coverage
