+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Add Admin note field to the Holdings field mapping profile: View (UIDATIMP-1122)
 * Add Admin note field to the Item field mapping profile: View (UIDATIMP-1123)
 * Display Authority information on Data import log page: View (UIDATIMP-1053)
+* Fix and Cover ViewAllLogs component with unit tests: Test (UIDATIMP-1151)
 
 ### Bugs fixed:
 * Data Import landing page log shows in old format instead of current format (UIDATIMP-1139)
